requirements.txt
Django~=4.2.3
django-ckeditor~=6.6.1
django-cleanup~=8.0.0
django-crispy-forms~=2.0
django-js-asset~=2.1.0
